reorganize tasks and added set and uset
